requirements.txt
langchain== 1.2.15
langchain-core ==1.3.2
langchain-google-genai== 4.2.2
openai ==2.33.0
pydantic ==2.13.3
python-dotenv== 1.2.2
pydantic== 2.13.3
